requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-genai>=1.0.0
httpx==0.28.1
boto3==1.35.86
python-multipart==0.0.20

